Fix linting issues + clean up
